// app/shared/model/trans.model.ts
export class Translate {

    constructor(){}

    id: number;
    title:string;
    created_date: Date;
    decription:string;
    dest_lang:number;
    destlang_name:string;
    end_date: Date;
    update_date: Date;
    translate_status: number;
    link_customer: string;
    link_translator: string;
    price: number;
    src_lang:number;
    srclang_name:string;
    status_name:string;
    id_customer: number;
    id_translator: number;
    id_type: number;
    typename: string;
    tr_userlname: string;
    tr_userfname: string;
    cs_userlname: string;
    cs_userfname: string;

}